requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
psycopg[binary]>=3.1.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client for external APIs
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security scanning
bandit==1.7.5

# Production dependencies

# Email validation for Pydantic
email-validator==2.1.0 